backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
av==16.0.1
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
ctranslate2==4.6.3
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
executing==2.2.1
fastapi==0.128.0
faster-whisper==1.2.1
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
filelock==3.20.2
flatbuffers==25.12.19
fsspec==2025.12.0
future==1.0.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.1
humanfriendly==10.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
mpmath==1.3.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.4.0
onnxruntime==1.23.2
openai==2.14.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.1.0
pipreqs==0.5.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
reportlab==4.4.7
requests==2.32.5
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tinycss2==1.4.0
tokenizers==0.22.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
yarg==0.1.9
